+++ static/js/projects.js
@@ -1,5 +1,6 @@
 $(function () {
     var base = document.getElementById("base-url").getAttribute("href").slice(0, -1);
+    var favicon = document.querySelector('link[rel="icon"]');
 
     // Schedule refresh to run every 3s when tab is visible and user
     // is active, every 60s otherwise
@@ -11,8 +12,10 @@ $(function () {
             dataType: "json",
             timeout: 2000,
             success: function(data) {
+                var anyDown = false;
                 for (var code in data) {
                     var el = data[code];
+                    anyDown = anyDown || (el.status == "down");
 
                     if (el.status != lastStatus[code]) {
                         $("#" + code + " div.status").attr("class", "status ic-" + el.status);
@@ -24,6 +27,8 @@ $(function () {
                         lastStarted[code] = el.started;
                     }
                 }
+                var downPostfix = anyDown ? "_down" : "";
+                favicon.href = `${base}/static/img/favicon${downPostfix}.svg`;
             }
         });
     }