+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 
 ### Improvements
 - Show status changes (flips) in check's log page (#447)
+- Implement dynamic favicon in the projects overview page (#971)
 
 ## v3.3 - 2024-04-03